--- src/main/java/toy/withme58/api/member/controller/MemberViewController.java
@@ -1,23 +1,0 @@
-package toy.withme58.api.member.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MemberViewController {
-
-    @GetMapping("/api/member/me")
-    public String showMyInfoPage(){
-        return "member/me";
-    }
-
-    @GetMapping("/open-api/member/register")
-    public String showRegisterPage(){
-        return "member/register";
-    }
-
-    @GetMapping("/open-api/member/login")
-    public String showLoginPage(){
-        return "member/login";
-    }
-}
