+++ src/main/java/toy/withme58/api/qustion/QuestionService.java
@@ -2,6 +2,7 @@ package toy.withme58.api.qustion;
 
 import org.springframework.stereotype.Service;
 import toy.withme58.db.question.QuestionEntity;
+import toy.withme58.db.question.QuestionRepository;
 
 import java.util.List;
