+++ src/main/java/toy/withme58/db/memberquestion/MemberQuestionEntity.java
@@ -1,0 +1,14 @@
+package toy.withme58.db.memberquestion;
+
+import jakarta.persistence.*;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@Table(name = "member_question")
+@Entity
+public class MemberQuestionEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+}
