+++ src/main/java/toy/withme58/db/question/QuestionEntity.java
@@ -1,10 +1,15 @@
 package toy.withme58.db.question;
 
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.NoArgsConstructor;
+import toy.withme58.db.question.enums.QuestionStatus;
 
 @Table(name = "question")
 @NoArgsConstructor
+@AllArgsConstructor
+@Builder
 @Entity
 public class QuestionEntity {
 
@@ -13,7 +18,9 @@ public class QuestionEntity {
     private Long id;
 
     @Column(nullable = false)
-    private String question;
+    private String title;
 
-
+    @Column(columnDefinition = "varchar(50)", nullable = false)
+    @Enumerated(EnumType.STRING)
+    private QuestionStatus status;
 }
