--- src/main/java/toy/withme58/db/question/QuestionMemberEntity.java
@@ -1,16 +1,0 @@
-package toy.withme58.db.question;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@Entity
-public class QuestionMemberEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-}
