+++ src/main/resources/application.yaml
@@ -1,4 +1,3 @@
-
 spring:
   jpa:
     show-sql: true
@@ -11,7 +10,7 @@ spring:
     url: jdbc:mysql://localhost:3306/yonombara?useSSL=false&useUnicode=true&allowPublicKeyRetrieval=true
     driver-class-name: com.mysql.cj.jdbc.Driver
     username: root
-    password: root
+    password: new_agnes0812
 
 token:
   secret:
